perf: add vss extension and incrementally load duckdb

+++ scripts/recipes/hybrid_search_duckdb.py
@@ -60,7 +60,7 @@ def main():
 				file_path,
 				md_content,
 				fts_main_vault.match_bm25(file_path, ?) AS full_text_score,
-				array_cosine_similarity(?::DOUBLE[1536], embeddings_openai) AS similarity
+				array_cosine_similarity(?::FLOAT[1536], embeddings_openai) AS similarity
 			FROM vault
 			WHERE 
 				embeddings_openai IS NOT NULL
